feat: make navbar responsive on all support
